Market framework added, protocols moved for use in trade as well as migration
