+++ .gitignore
@@ -1,0 +1,37 @@
+__pycache__/config.cpython-311.pyc
+__pycache__/main.cpython-311.pyc
+model/city/__pycache__/__init__.cpython-311.pyc
+model/city/__pycache__/city.cpython-311.pyc
+model/city/__pycache__/city_data.cpython-311.pyc
+model/core/__pycache__/__init__.cpython-311.pyc
+model/core/__pycache__/ticks.cpython-311.pyc
+model/migration/__pycache__/migration.cpython-311.pyc
+model/migration/__pycache__/__init__.cpython-311.pyc
+model/population/__pycache__/group.cpython-311.pyc
+model/population/__pycache__/__init__.cpython-311.pyc
+model/province/__pycache__/province.cpython-311.pyc
+model/province/__pycache__/__init__.cpython-311.pyc
+visualisation/__pycache__/report.cpython-311.pyc
+model/country/__pycache__/__init__.cpython-311.pyc
+model/country/__pycache__/country.cpython-311.pyc
+model/core/__pycache__/random.cpython-311.pyc
+visualisation/__pycache__/graph.cpython-311.pyc
+model/economy/__pycache__/__init__.cpython-311.pyc
+model/economy/industry/__pycache__/__init__.cpython-311.pyc
+model/economy/labour/__pycache__/labour_market.cpython-311.pyc
+model/economy/labour/__pycache__/__init__.cpython-311.pyc
+model/economy/industry/__pycache__/firm.cpython-311.pyc
+model/city/__pycache__/labour_market.cpython-311.pyc
+tests/__pycache__/test_sim_regressions.cpython-311.pyc
+model/core/__pycache__/invariants.cpython-311.pyc
+.vscode/settings.json
+model/province/province.py
+tests/__pycache__/test_migration_phase1.cpython-311.pyc
+model/migration/__pycache__/allocation.cpython-311.pyc
+model/migration/__pycache__
+model/migration/engines/__pycache__
+tests/__pycache__
+model/__pycache__/protocols.cpython-311.pyc
+model/__pycache__/__init__.cpython-311.pyc
+model/economy/trade/__pycache__/__init__.cpython-311.pyc
+model/economy/trade/__pycache__/market.cpython-311.pyc
