+++ model/__init__.py
@@ -1,0 +1,1 @@
+from .protocols import DistanceProvider, NeutralDistanceProvider
