+++ model/core/ticks.py
@@ -19,6 +19,8 @@ class Core:
         self.province_cfg = province_cfg
         self.country_cfg = country_cfg
 
+        self.countries = []
+
     def tick(self):
         for country in self.countries:
             country.tick()
@@ -50,8 +52,6 @@ class Core:
         return [self._build_province(province_data) for province_data in data["provinces"]]
 
     def build_sim(self):
-        self.countries = []
-
         with open("input_data.json") as f:
             data = json.load(f)
 
@@ -63,5 +63,5 @@ class Core:
             country_obj = Country(provinces, name, cfg, rng=self.rng)
 
             self.countries.append(country_obj)
-        
-    
+
+
