+++ model/economy/__init__.py
@@ -1,2 +1,3 @@
 from .industry.firm import Firm
 from .labour.labour_market import LabourMarket
+from .trade.market import Market
