+++ model/economy/trade/__init__.py
@@ -1,0 +1,1 @@
+from .market import Market
