+++ model/migration/__init__.py
@@ -1,13 +1,10 @@
 from .facade import Migration
 from .options import MigrationOptions, MigrationConfig
-from .protocols import DistanceProvider, NeutralDistanceProvider
 from .types import GroupMigrationEvent
 
 __all__ = [
-    "DistanceProvider",
     "GroupMigrationEvent",
     "Migration",
     "MigrationOptions",
     "MigrationConfig",
-    "NeutralDistanceProvider",
 ]
